+++ edgy/core/db/querysets/compiler.py
@@ -6,7 +6,6 @@ from typing import TYPE_CHECKING, Any
 import sqlalchemy
 
 from edgy.core.db.fields.base import BaseForeignKey, RelationshipField
-from edgy.core.db.relationships.utils import crawl_relationship
 from edgy.core.utils.db import get_table_key_or_name, hash_tablekey
 from edgy.exceptions import QuerySetError
 
@@ -317,13 +316,9 @@ class QueryCompiler:
         _select_tables_and_models: tables_and_models_type = {"": (select_from, self.model_class)}
         transitions: dict[tuple[str, str, str], tuple[Any, tuple[str, str, str] | None, str]] = {}
 
-        select_pathes = self.queryset._select_related.union(self.queryset._select_related_weak)
-        if self.queryset.embed_parent:
-            # implicit select the forward path of embed_parent
-            result = crawl_relationship(
-                self.queryset.model_class, self.queryset.embed_parent[0], traverse_last=True
-            )
-            select_pathes.add(result.forward_path)
+        select_pathes = self.queryset._select_related.union(
+            self.queryset._select_related_g_and_o, self.queryset._select_related_embedding
+        )
 
         for select_path in select_pathes:
             model_class = self.model_class